remove hunyuan_common.py
